+++ es5/tools/debug.js
@@ -32,6 +32,7 @@
     exports.default = Debug;
     /**
      * LVL99 Debug
+     *
      * A console-like replacement which creates a noop console object if you don't want to output stuff via the console
      */