+++ es5/tools/index.js
@@ -1,29 +1,29 @@
 (function (global, factory) {
   if (typeof define === "function" && define.amd) {
-    define(['module', 'exports', './breakpoints', './debug', './queue', './trackevent', './smooth-scroll'], factory);
+    define(['module', 'exports', './breakpoints', './debug', './queue', './trackevent', './smooth-scroll', './storage'], factory);
   } else if (typeof exports !== "undefined") {
-    factory(module, exports, require('./breakpoints'), require('./debug'), require('./queue'), require('./trackevent'), require('./smooth-scroll'));
+    factory(module, exports, require('./breakpoints'), require('./debug'), require('./queue'), require('./trackevent'), require('./smooth-scroll'), require('./storage'));
   } else {
     var mod = {
       exports: {}
     };
-    factory(mod, mod.exports, global.breakpoints, global.debug, global.queue, global.trackevent, global.smoothScroll);
+    factory(mod, mod.exports, global.breakpoints, global.debug, global.queue, global.trackevent, global.smoothScroll, global.storage);
     global.index = mod.exports;
   }
 })(this, function (module, exports) {
   (function (global, factory) {
     if (typeof define === "function" && define.amd) {
-      define(['module', 'exports', './breakpoints', './debug', './queue', './trackevent', './smooth-scroll'], factory);
+      define(['module', 'exports', './breakpoints', './debug', './queue', './trackevent', './smooth-scroll', './storage'], factory);
     } else if (typeof exports !== "undefined") {
       factory(module, exports);
     } else {
       var mod = {
         exports: {}
       };
-      factory(mod, mod.exports, global.breakpoints, global.debug, global.queue, global.trackevent, global.smoothScroll);
+      factory(mod, mod.exports, global.breakpoints, global.debug, global.queue, global.trackevent, global.smoothScroll, global.storage);
       global.index = mod.exports;
     }
-  })(this, function (module, exports, _breakpoints, _debug, _queue, _trackevent, _smoothScroll) {
+  })(this, function (module, exports, _breakpoints, _debug, _queue, _trackevent, _smoothScroll, _storage) {
     'use strict';
 
     Object.defineProperty(exports, "__esModule", {
@@ -40,23 +40,28 @@
 
     var _smoothScroll2 = _interopRequireDefault(_smoothScroll);
 
+    var _storage2 = _interopRequireDefault(_storage);
+
     function _interopRequireDefault(obj) {
       return obj && obj.__esModule ? obj : {
         default: obj
       };
     }
 
+    /**
+     * LVL99 Tools
+     *
+     * Standalone tools that don't require any dependencies within the framework, but work alongside
+     */
+
     var utils = {
       Breakpoints: _breakpoints2.default,
       Debug: _debug2.default,
       Queue: _queue2.default,
       TrackEvent: _trackevent2.default,
-      SmoothScroll: _smoothScroll2.default
-    }; /**
-        * LVL99 Tools
-        *
-        * Standalone tools that don't require any dependencies within the framework, but work alongside
-        */
+      SmoothScroll: _smoothScroll2.default,
+      Storage: _storage2.default
+    };
 
     exports.default = tools;
     module.exports = exports['default'];