+++ es5/utils/parse.js
@@ -63,11 +63,13 @@
      * @returns {Mixed}
      */
     function coerceToPrimitiveType(input) {
-      // Non-string? Just return it straight away
-      if (typeof input !== 'string') return input;
+      // Non-string or empty string? Just return it straight away
+      if (typeof input !== 'string' || input === '') {
+        return input;
+      }
 
       // Trim any whitespace
-      input = (input + '').trim();
+      var output = (input + '').trim();
 
       // Number
       if (/^\-?(?:\d*[\.\,])*\d*(?:[eE](?:\-?\d+)?)?$/.test(input)) {
@@ -156,7 +158,7 @@
         try {
           output = JSON.parse(input);
         } catch (e) {
-          console.error(_loggerPath + '.convertStringToJson: Error parsing string JSON data', input);
+          console.error(__loggerPath + '.convertStringToJson: Error parsing string JSON data', input);
         }
       }
 
@@ -297,15 +299,15 @@
 
       // No object found!
       if ((typeof trigger === 'undefined' ? 'undefined' : _typeof(trigger)) !== 'object') {
-        throw new Error(_loggerPath + '.extractTriggerDetails: input was not valid JSON or CSS-style definition');
+        throw new Error(__loggerPath + '.extractTriggerDetails: input was not valid JSON or CSS-style definition');
       }
 
       // Ensure it has `on` and `do` properties
       // if (!objectPath.has(trigger, 'on')) {
-      //   throw new Error(`${_loggerPath}.extractTriggerDetails: trigger is missing required 'on' property`)
+      //   throw new Error(`${__loggerPath}.extractTriggerDetails: trigger is missing required 'on' property`)
       // }
       if (!_objectPath2.default.has(trigger, 'do')) {
-        throw new Error(_loggerPath + '.extractTriggerDetails: trigger is missing required \'do\' property');
+        throw new Error(__loggerPath + '.extractTriggerDetails: trigger is missing required \'do\' property');
       }
 
       // If target is set, use real values for window and document