+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 * Added new tool [`Storage`](es6/tools/storage.es6) which acts as a convenience wrapper with some extra small features
   when using local/session storage.
+* Added tests for [`Storage`](__tests__/tools/storage.test.es6) and [`SmoothScroll`](__tests__/tools/smooth-scroll.test.es6)
+  tools.
+* Refactored some of the configuration setup for Babel and Jest.
 
 
 ### 1.4.2